+++ classes/Shipper.class.php
@@ -23,6 +23,9 @@ class Shipper
 {
     use DBO;
 
+    const TAX_DESTINATION = 1;
+    const TAX_ORIGIN = 0;
+
     /** Table key for DBO functions
      * @var string */
     private static $TABLE = 'shop.shipping';
@@ -49,7 +52,7 @@ class Shipper
 
     /** glFusion group ID which can use this shipper.
      * @var integer */
-    private $grp_access = 1;
+    private $grp_access = 2;
 
     /** Name of the shipper, e.g. "United Parcel Service".
      * @var string */
@@ -73,7 +76,7 @@ class Shipper
 
     /** Flag to indicate whether fixed shipping cost is included or ignored.
      * @var boolean */
-    private $use_fixed = false;
+    private $use_fixed = 1;
 
     /** Std class to accumulate order shipping prices for the shipper.
      * This is manipulated as a public variable within this class.
@@ -116,6 +119,16 @@ class Shipper
      * @var string */
     protected $key = 'generic';
 
+    /** Flag to indicate that a shipping address is required.
+     * For "will-call" or certain other methods an address may not be needed.
+     * @var boolean */
+    protected $req_shipto = true;
+
+    /** Flag to indicate where sales tax is calculated.
+     * 1 = origin (shop address), 0 = destination (customer address)
+     * @var integer */
+    protected $tax_loc = 0;
+
 
     /**
      * Constructor. Sets variables from the provided array.
@@ -136,14 +149,8 @@ class Shipper
             if ($this->Read($A)) $this->isNew = false;
         } else {
             // New entry, set defaults
-            $this->id = 0;
-            $this->module_code = '';
-            $this->enabled = 1;
-            $this->name = '';
-            $this->use_fixed = 1;
             $this->setValidFrom(NULL);
             $this->setValidTo(NULL);
-            $this->grp_access = 2;    // Default = All users
             $this->min_units = self::MIN_UNITS;
             $this->max_units = 1000000;
             $this->rates = array(
@@ -172,9 +179,9 @@ class Shipper
         global $_TABLES;
 
         $id = (int)$id;
-        $cache_key = self::$base_tag . ' _ ' . $id;
-        $A = Cache::get($cache_key);
-        if ($A === NULL) {
+        //$cache_key = self::$base_tag . ' _ ' . $id;
+        //$A = Cache::get($cache_key);
+        //if ($A === NULL) {
             $sql = "SELECT *
                     FROM {$_TABLES['shop.shipping']}
                     WHERE id = $id";
@@ -182,9 +189,9 @@ class Shipper
             $res = DB_query($sql);
             if ($res) {
                 $A = DB_fetchArray($res, false);
-                Cache::set($cache_key, $A, self::$base_tag);
+          //      Cache::set($cache_key, $A, self::$base_tag);
             }
-        }
+        //}
         if (!empty($A)) {
             $this->setVars($A);
             return true;
@@ -210,6 +217,8 @@ class Shipper
             ->setMinUnits(SHOP_getVar($A, 'min_units', 'integer'))
             ->setMaxUnits(SHOP_getVar($A, 'max_units', 'integer'))
             ->setEnabled(SHOP_getVar($A, 'enabled', 'integer'))
+            ->setReqShipto(SHOP_getVar($A, 'req_shipto', 'integer'))
+            ->setTaxLocation(SHOP_getVar($A, 'tax_loc', 'integer'))
             ->setUseFixed(SHOP_getVar($A, 'use_fixed', 'integer', 0))
             ->setGrpAccess(SHOP_getVar($A, 'grp_access', 'integer', 2));
         if (!$fromDB) {
@@ -360,6 +369,44 @@ class Shipper
         $this->enabled = $enabled ? 1 : 0;
         return $this;
     }
+
+
+    /**
+     * Set the `require shipto address` flag value.
+     *
+     * @param   boolean $enabled    True or false
+     * @return  object  $this
+     */
+    private function setReqShipto($flag)
+    {
+        $this->req_shipto = $flag ? 1 : 0;
+        return $this;
+    }
+
+
+    /**
+     * Get the value of the "requires shipto" flag.
+     *
+     * @return  integer     1 if required, 0 if not
+     */
+    public function requiresShipto()
+    {
+        return $this->req_shipto ? 1 : 0;
+    }
+
+
+    private function setTaxLocation($flag)
+    {
+        $this->tax_loc = $flag ? 1 : 0;
+        return $this;
+    }
+
+
+    public function getTaxLocation()
+    {
+        return (int)$this->tax_loc;
+    }
+
 
     /**
      * Set the flag to include product fixed shipping charge.
@@ -843,6 +890,8 @@ class Shipper
         $sql2 = " SET name = '" . DB_escapeString($this->name) . "',
             module_code= '" . DB_escapeString($this->module_code) . "',
             enabled = '{$this->enabled}',
+            req_shipto = '{$this->requiresShipto()}',
+            tax_loc = '{$this->getTaxLocation()}',
             min_units = '{$this->min_units}',
             max_units = '{$this->max_units}',
             valid_from = '{$this->valid_from->toUnix()}',
@@ -910,10 +959,12 @@ class Shipper
             'min_units'     => $this->min_units == self::MIN_UNITS ? 0 : $this->min_units,
             'max_units'     => $this->max_units,
             'ena_sel'       => $this->enabled ? 'checked="checked"' : '',
+            'req_shipto_sel' => $this->req_shipto ? 'checked="checked"' : '',
             'fixed_sel'     => $this->use_fixed ? 'checked="checked"' : '',
             'valid_from'    => $this->valid_from->format('Y-m-d', true),
             'valid_to'      => $this->valid_to->format('Y-m-d', true),
             'grp_sel'       => COM_optionList($_TABLES['groups'], 'grp_id,grp_name', $this->grp_access),
+            'tax_loc_' . $this->getTaxLocation() => 'selected="selected"',
         ) );