+++ language/english_utf-8.php
@@ -712,6 +712,9 @@ $LANG_SHOP = array (
 'bal_due' => 'Balance Due',
 'notify_pmt_received' => 'Your payment has been received.',
 'dup_pxf' => 'Duplicate or invalid feature selected.',
+'tax_loc' => 'Sales Tax Location',
+'origin' => 'Origin',
+'destination' => 'Destination',
 );
 if (isset($_SHOP_CONF['ena_ratings']) && $_SHOP_CONF['ena_ratings']) {
     $LANG_SHOP['list_sort_options']['top_rated'] = 'Top Rated';